+++ ETS2LA/ReleaseNotes.md
@@ -1,4 +1,5 @@
-### ETS2LA C# 3.1.4
-* Fix `GLFW Error 65544` spam on some machines.
+### ETS2LA C# 3.1.5
+* `ETS2LA.Game` now supports Linux, although it assumes all games are the latest version (1.58.0) due to the lack of file version info on Linux.
+* `ETS2LA.Controls` now uses device GUIDs instead of instance IDs. This means that binds survive past restarts and device re-plugging.
 
 **NOTE:** If you're not in the ETS2LA beta program, please take a look at https://ets2la.com/downloads. This version is not what you're looking for!